Create Q1.cpp
Round Robin is the pre-emptive process scheduling algorithm. Each process is provided a fix time to execute, it is called a quantum. Once a process is executed for a given time period, it is preempted and other process executes for a given time period. Context switching is used to save states of preempted processes.

In this program, I have simulated the Round robin algorithm with sample inputs. value of quantum = 3